Fixed rss output glitch, added ad filter
Added filter to exclude wallflux rss feeds that include "wallflux" wihtin the link for the RSS entry.

Wall flux is great but the ads seem to become more common for feeds with less activity. It tends to also bring up older posts which is something that'll have to be addressed too with a date check function.